+++ instrumentation/opentelemetry-instrumentation-pyramid/tests/test_automatic.py
@@ -17,6 +17,7 @@ from pyramid.config import Configurator
 from opentelemetry.instrumentation.pyramid import PyramidInstrumentor
 from opentelemetry.test.test_base import TestBase
 from opentelemetry.test.wsgitestutil import WsgiTestBase
+from opentelemetry.trace import SpanKind
 
 # pylint: disable=import-error
 from .pyramid_base_test import InstrumentationTest
@@ -77,3 +78,34 @@ class TestAutomatic(InstrumentationTest, TestBase, WsgiTestBase):
         self.assertEqual([b"Hello: 123"], list(resp.response))
         span_list = self.memory_exporter.get_finished_spans()
         self.assertEqual(len(span_list), 1)
+
+
+class TestWrappedWithOtherFramework(
+    InstrumentationTest, TestBase, WsgiTestBase
+):
+    def setUp(self):
+        super().setUp()
+        PyramidInstrumentor().instrument()
+        self.config = Configurator()
+        self._common_initialization(self.config)
+
+    def tearDown(self) -> None:
+        super().tearDown()
+        with self.disable_logging():
+            PyramidInstrumentor().uninstrument()
+
+    def test_with_existing_span(self):
+        tracer_provider, _ = self.create_tracer_provider()
+        tracer = tracer_provider.get_tracer(__name__)
+
+        with tracer.start_as_current_span(
+            "test", kind=SpanKind.SERVER
+        ) as parent_span:
+            resp = self.client.get("/hello/123")
+            self.assertEqual(200, resp.status_code)
+            span_list = self.memory_exporter.get_finished_spans()
+            self.assertEqual(SpanKind.INTERNAL, span_list[0].kind)
+            self.assertEqual(
+                parent_span.get_span_context().span_id,
+                span_list[0].parent.span_id,
+            )